fix: repair v1 screens broken by shell quarantine move (#50)
git mv로 src/shells -> src/shells-v1 격리(#50)했을 때, 6개 v1 화면이 여전히
@/shells/ManagerShell(별칭 경로)를 참조하고 있어서 새로 만드는 v2
ConsoleShell에 잘못 바인딩됐다 — 별칭은 파일이 이동해도 자동으로 안 따라온다.
전부 @/shells-v1/ManagerShell로 고쳐 원래 v1 셸을 다시 가리키게 했다.

Wordmark는 인증 화면(BrandPanel) 전용이 아니라 콘솔 셸 헤더도 같은 로고를
쓰므로 components/common/으로 승격했다 — v1 안에 갇혀 있으면 v2에서 재사용할
때 v1 트리에 손을 대야 한다.
