+++ src/features-v1/curriculum/CurriculumUploadScreen.tsx
@@ -2,7 +2,7 @@ import { useRef, useState, type DragEvent } from 'react'
 import { Controller, useForm } from 'react-hook-form'
 import { Link, useNavigate } from 'react-router'
 import { ArrowLeft, FileIcon, InfoIcon, UploadIcon, XIcon } from 'lucide-react'
-import ManagerShell from '@/shells/ManagerShell'
+import ManagerShell from '@/shells-v1/ManagerShell'
 import PageHeader from '@/components/common/PageHeader'
 import { Badge } from '@/components/ui/Badge'
 import { Button } from '@/components/ui/Button'
